django_blog/requirements.txt
psycopg2-binary
psycopg2
Django
Pillow
django-bootstrap4
django-crispy-forms
django-extensions
persiantools
djangorestframework
